[meta] slight reorganization of reporter
Move JUnit stuff to being a submodule of the aggregator, and follow the
mod.rs/imp.rs pattern.
