+++ nextest-runner/src/reporter/aggregator/mod.rs
@@ -1,0 +1,9 @@
+// Copyright (c) The nextest Contributors
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+//! Aggregates events, storing them in the store directory.
+
+mod imp;
+mod junit;
+
+pub(crate) use imp::*;
